Check api (#17)
* add customer_order and fix sth

* add_api_Content_Promotion

* add_Redeem_Promotion

* fix conflict1

* fix conflict2

* fix conflict3

* fix conflict3

* fix conflict4

* fix conflict5

* fix conflict4

* fix conflict5

* fix conflict9

* fix conflict10

---------

+++ WebSiteBanHang/Repositories/EFProductRepository.cs
@@ -12,17 +12,11 @@ namespace WebSiteBanHang.Repositories
         }
         public async Task<IEnumerable<Product>> GetAllAsync()
         {
-            // return await _context.Products.ToListAsync();
-            return await _context.Products
-            .Include(p => p.Category) // Include thông tin về category
-            .ToListAsync();
+            return await _context.Products.ToListAsync();
         }
-        public async Task<Product> GetByIdAsync(int id)
+        public async Task<Product?> GetByIdAsync(int id)
         {
-            // return await _context.Products.FindAsync(id);
-            // lấy thông tin kèm theo category
-            return await _context.Products.Include(p =>
-           p.Category).FirstOrDefaultAsync(p => p.Id == id);
+            return await _context.Products.FindAsync(id);
         }
         public async Task AddAsync(Product product)
         {
@@ -31,14 +25,25 @@ namespace WebSiteBanHang.Repositories
         }
         public async Task UpdateAsync(Product product)
         {
-            _context.Products.Update(product);
-            await _context.SaveChangesAsync();
+            var existingProduct = await _context.Products.FindAsync(product.Id);
+            if (existingProduct != null)
+            {
+                _context.Entry(existingProduct).CurrentValues.SetValues(product);
+                await _context.SaveChangesAsync();
+            }
         }
         public async Task DeleteAsync(int id)
         {
             var product = await _context.Products.FindAsync(id);
-            _context.Products.Remove(product);
-            await _context.SaveChangesAsync();
+            if (product != null)
+            {
+                _context.Products.Remove(product);
+                await _context.SaveChangesAsync();
+            }
+        }
+        public async Task<int> GetProductCountAsync()
+        {
+            return await _context.Products.CountAsync();
         }
         public async Task<int> GetProductCountAsync()
         {
